pyproject.toml
[project]
name = "bitovi-project"
version = "0.1.0"
description = "AI agent with RAG layer and UI that ingests blog data and answers questions about it"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "beautifulsoup4",
    "einops",
    "fastapi",
    "langchain<1",
    "langchain-anthropic",
    "langchain-chroma",
    "langchain-community",
    "langchain-huggingface",
    "langchain-text-splitters",
    "lxml",
    "opentelemetry-exporter-otlp-proto-grpc>=1.41.0",
    "python-dotenv",
    "sentence-transformers",
    "tiktoken",
    "uvicorn",
]

[dependency-groups]
dev = [
    "pytest",
]

[tool.uv.workspace]
members = [
    "frontend",
]
